+++ tests/Instances/Data/Semigroup.hs
@@ -1,6 +1,3 @@
-{-# LANGUAGE CPP                        #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE StandaloneDeriving         #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 {-|
@@ -11,21 +8,13 @@ Maintainer:  Ryan Scott
 Stability:   Provisional
 Portability: GHC
 
-'Arbitrary' instances for datatypes in the "Data.Semigroup" module.
+'Arbitrary' instance for the 'Arg' datatype.
 -}
 module Instances.Data.Semigroup () where
 
-import Data.Semigroup (Min(..), Max(..), First(..), Last(..),
-                       WrappedMonoid(..), Option(..), Arg(..))
+import Data.Semigroup (Arg(..))
 import Instances.Utils.GenericArbitrary (genericArbitrary)
 import Test.QuickCheck (Arbitrary(..))
-
-deriving instance Arbitrary a => Arbitrary (Min a)
-deriving instance Arbitrary a => Arbitrary (Max a)
-deriving instance Arbitrary a => Arbitrary (First a)
-deriving instance Arbitrary a => Arbitrary (Last a)
-deriving instance Arbitrary a => Arbitrary (WrappedMonoid a)
-deriving instance Arbitrary a => Arbitrary (Option a)
 
 instance (Arbitrary a, Arbitrary b) => Arbitrary (Arg a b) where
     arbitrary = genericArbitrary